return 404 when deleting a org with non-exist id

+++ src/main/java/edu/sjsu/cmpe275/lab2/dao/impl/HibernateOrganizationDao.java
@@ -66,6 +66,9 @@ public class HibernateOrganizationDao implements OrganizationDao {
         try{
             tx.begin();
             Organization org = session.get(Organization.class, orgId);
+            if (org == null) {
+                throw new RuntimeException();
+            }
             if (org.getPersons().isEmpty()) {
                 session.delete(org);
             } else {
